requirements.txt
flask==2.0.1
scikit-learn==0.24.2
pandas==1.3.0
numpy==1.21.0
joblib==1.0.1
python-dotenv==0.19.0
sqlalchemy==1.4.23
flask-cors==3.0.10
gunicorn==20.1.0